+++ Auditor/Modules/Preprocessor/PreprocessorModule.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Reflection;
+using System.Runtime.InteropServices.WindowsRuntime;
 using NUnit.Framework;
 using UnityEditor;
 using UnityEngine;
@@ -11,22 +12,24 @@ namespace AssetTools
 	[Serializable]
 	public class PreprocessorModule : IImportProcessModule
 	{
-		private static Assembly[] m_Assemblies;
-
-		
 		/// <summary>
 		/// TODO If any of this is changed, do the Assets imported by it need to be reimported?
 		/// </summary>
 		
-		private string m_Method;
-		public string m_Data;
+		[SerializeField] private string m_MethodString;
+		[SerializeField] private string m_Data;
 
 		// used for locking it to a particular asset type
 		public string m_SearchFilter;
-		
+
 		
 		private List<string> m_AssetsToForceApply = new List<string>();
-		
+
+		public string methodString
+		{
+			get { return m_MethodString; }
+		}
+
 		public bool CanProcess( AssetImporter item )
 		{
 			return true;
@@ -37,12 +40,6 @@ namespace AssetTools
 			return m_AssetsToForceApply.Contains( item.assetPath );
 		}
 		
-		
-		[InitializeOnLoadMethod]
-		static void CollectAssemblies()
-		{
-			m_Assemblies = AppDomain.CurrentDomain.GetAssemblies();
-		}
 		
 		public List<IConformObject> GetConformObjects( string asset )
 		{
@@ -75,14 +72,8 @@ namespace AssetTools
 			AssetViewItem selectedNodes = context as AssetViewItem;
 			if( selectedNodes != null )
 			{
-				// Set the userData so the import pipeline knows that it should be m_Importer with the method data
-				SetUserData( selectedNodes.AssetImporter );
+				m_AssetsToForceApply.Add( selectedNodes.path );
 				selectedNodes.ReimportAsset();
-				
-				// if( Apply( selectedNodes.AssetImporter ) )
-				// {
-				// 	selectedNodes.ReimportAsset();
-				// }
 				
 				foreach( IConformObject data in selectedNodes.conformData )
 				{
@@ -106,21 +97,24 @@ namespace AssetTools
 				Debug.LogError( "Could not fix Asset with no Assets selected." );
 		}
 
-		private void SetUserData( AssetImporter importer )
+		private void SetUserData( AssetImporter importer, AuditProfile profile )
 		{
-			// TODO this
+			UserDataSerialization data = UserDataSerialization.Get( importer.assetPath );
+			string profileGuid = AssetDatabase.AssetPathToGUID( AssetDatabase.GetAssetPath( profile ) );
+			data.m_ImporterPostprocessorData.UpdateOrAdd( new UserDataSerialization.PostprocessorData( profileGuid, "PreprocessorModule", Method.AssemblyName, Method.TypeName, Method.Version ) );
+			data.UpdateImporterUserData();
 		}
 		
-		public bool Apply( AssetImporter item )
+		public bool Apply( AssetImporter item, AuditProfile fromProfile )
 		{
-			if( string.IsNullOrEmpty( m_Method ) == false )
+			if( string.IsNullOrEmpty( m_MethodString ) == false )
 			{
-				MethodInfo postprocessorMethodToInvoke = GetMethodInfo( m_Method );
-				if( postprocessorMethodToInvoke != null )
+				if( Method != null )
 				{
-					object returnValue = postprocessorMethodToInvoke.Invoke( null, string.IsNullOrEmpty( m_Data ) ? null : new object[] {m_Data} );
+					object returnValue = Method.Invoke( item, m_Data );
 					if( returnValue != null )
 					{
+						SetUserData( item, fromProfile );
 						return (bool) returnValue;
 					}
 				}
@@ -128,43 +122,109 @@ namespace AssetTools
 			return false;
 		}
 
-		private static MethodInfo GetMethodInfo( string m_Method )
+		internal ProcessorMethodInfo m_ProcessorMethodInfo;
+
+		private ProcessorMethodInfo Method
+		{
+			get
+			{
+				if( m_ProcessorMethodInfo == null && string.IsNullOrEmpty( m_MethodString ) == false )
+				{
+					string assemblyName;
+					string typeString;
+					GetMethodStrings( out assemblyName, out typeString );
+					if( string.IsNullOrEmpty( typeString ) )
+					{
+						Debug.LogError( "Error collecting method from " + m_MethodString );
+						return null;
+					}
+					
+					List<ProcessorMethodInfo> methods = PreprocessorImplementorCache.Methods;
+					for( int i = 0; i < methods.Count; ++i )
+					{
+						if( assemblyName != null && methods[i].AssemblyName.StartsWith( assemblyName ) == false )
+							continue;
+
+						if( methods[i].TypeName.EndsWith( typeString ) )
+						{
+							m_ProcessorMethodInfo = methods[i];
+							break;
+						}
+					}
+				}
+		
+				return m_ProcessorMethodInfo;
+			}
+		}
+		
+		public void GetMethodStrings( out string assemblyName, out string typeString )
+		{
+			int commaIndex = m_MethodString.IndexOf( ',' );
+			if( commaIndex > 0 )
+			{
+				assemblyName = m_MethodString.Substring( commaIndex + 2 );
+				typeString = m_MethodString.Substring( 0, commaIndex );
+			}
+			else
+			{
+				assemblyName = "";
+				typeString = m_MethodString;
+			}
+		}
+		
+		/*
+		 
+		 // DECISION Would it be better to get the method at Apply time?
+		
+		internal MethodInfo m_MethodInfo = null;
+		private static Assembly[] m_Assemblies;
+		[InitializeOnLoadMethod]
+		static void CollectAssemblies()
+		{
+			m_Assemblies = AppDomain.CurrentDomain.GetAssemblies();
+		}
+
+
+		private static MethodInfo GetMethodInfo( string m_MethodInfoString )
 		{
 			Assembly selected = null;
 			string typeString;
-			string methodName;
-			int commaIndex = m_Method.LastIndexOf( ',' );
+			string typeName;
+			int commaIndex = m_MethodInfoString.IndexOf( ',' );
 			
 			if( commaIndex > 0 )
 			{
-				string assemblyName = m_Method.Substring( commaIndex + 2 );
+				string assemblyName = m_MethodInfoString.Substring( commaIndex + 2 );
 				// has an Assembly defined in the string. Use that
 				for( int i = 0; i < m_Assemblies.Length; ++i )
 				{
-					if( m_Assemblies[i].GetName().Name != assemblyName )
+					if( m_Assemblies[i].FullName != assemblyName )
 						continue;
 					selected = m_Assemblies[i];
 					break;
 				}
 
+				string methodInfoString = m_MethodInfoString.Substring( 0, commaIndex ) + ".OnPreprocessAsset";
+
 				int lastStop = 0;
 				for( int index = commaIndex; index >= 0; --index )
 				{
-					if( m_Method[index] == '.' )
+					if( methodInfoString[index] == '.' )
 					{
 						lastStop = index;
 						break;
 					}
 				}
 
-				Assert.IsTrue( lastStop > 0, "Error: Invalid Method string." );
-				methodName = m_Method.Substring( lastStop+1, (commaIndex-lastStop) - 1 );
-				typeString = m_Method.Substring( 0, lastStop );
+				Assert.IsTrue( lastStop > 0, "Error: Invalid methodString string." );
+				typeName = methodInfoString.Substring( lastStop+1 );
+				typeString = methodInfoString.Substring( 0, lastStop );
 			}
 			else
 			{
-				methodName = m_Method.Substring( m_Method.LastIndexOf( '.' ) + 1 );
-				typeString = m_Method.Substring( 0, m_Method.LastIndexOf( '.' ) );
+				m_MethodInfoString = m_MethodInfoString + ".OnPreprocessAsset";
+				typeName = m_MethodInfoString.Substring( m_MethodInfoString.LastIndexOf( '.' ) + 1 );
+				typeString = m_MethodInfoString.Substring( 0, m_MethodInfoString.LastIndexOf( '.' ) );
 			}
 
 			Type reflectedType = null;
@@ -184,16 +244,19 @@ namespace AssetTools
 
 			if( reflectedType == null )
 			{
-				Debug.LogWarning( string.Format( "Invalid method address for PostProcessMethod for {0}. Could not find type", "_______________________" ) );
+				Debug.LogWarning( string.Format( "Invalid method address for PostProcessMethod for {0}. Could not find type", m_MethodInfoString ) );
 				return null;
 			}
 
-			MethodInfo postprocessorMethodToInvoke = reflectedType.GetMethod( methodName, BindingFlags.Static | BindingFlags.Public );
+			// will always be public void OnPreprocessAsset( AssetImporter importer, string data )
+			MethodInfo postprocessorMethodToInvoke = reflectedType.GetMethod( typeName, BindingFlags.Public | BindingFlags.Instance, null, CallingConventions.Any, new []{ typeof(AssetImporter), typeof(string) }, null );
 
 			if( postprocessorMethodToInvoke == null )
-				Debug.LogWarning( string.Format( "Invalid method address for PostProcessMethod for {0}. Could not find method", "_____________________" ) );
+				Debug.LogWarning( string.Format( "Invalid method address for PostProcessMethod for {0}. Could not find method", m_MethodInfoString ) );
 
 			return postprocessorMethodToInvoke;
 		}
+		
+		*/
 	}
 }