+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # Changelog
 All notable changes will be documented in this file.
 
-## [1.0.1] - 2018-11-07
+## [0.0.3] - 2019-04-01
+- Added Preprocessor module
+- Added Profile sorting
+- Added Lock to directory for Profile filters
+
+## [0.0.2] - 2018-11-07
 - Removed ImporterType from list of filter Targets
 - Added labels as a filter Target
 - Filters are matched by OrdinalIgnoreCase where possible
 
-## [1.0.0] - 2018-11-06
+## [0.0.1] - 2018-11-06
 - Initial submission